ci: build one GraalVM JSON image
