+++ ci/run_ci.sh
@@ -79,8 +79,8 @@ install_jdks() {
   done
 }
 
-run_graalvm_tests() {
-  local main_classes=("$@")
+run_graalvm_test() {
+  local main_class="$1"
   local java_version
   local java_major
   java_version=$(java -version 2>&1 | awk -F '"' '/version/ {print $2; exit}')
@@ -103,25 +103,21 @@ run_graalvm_tests() {
     -Dmaven.source.skip=true \
     -Dmaven.javadoc.skip=true
   cd "$ROOT"/integration_tests/graalvm_tests
-  for main_class in "${main_classes[@]}"; do
-    echo "Start to build GraalVM JPMS native image for $main_class"
-    mvn -DmainClass="$main_class" -DskipTests=true -Dassembly.skipAssembly=true \
-      --no-transfer-progress -Pnative-module clean package
-    echo "Built GraalVM JPMS native image"
-    echo "Start to run GraalVM JPMS native image"
-    ./target/main-module
-    echo "Execute GraalVM tests for $main_class succeed!"
-  done
+  echo "Start to build GraalVM JPMS native image for $main_class"
+  mvn -DmainClass="$main_class" -DskipTests=true -Dassembly.skipAssembly=true \
+    --no-transfer-progress -Pnative-module clean package
+  echo "Built GraalVM JPMS native image"
+  echo "Start to run GraalVM JPMS native image"
+  ./target/main-module
+  echo "Execute GraalVM tests for $main_class succeed!"
 }
 
 graalvm_test() {
-  run_graalvm_tests org.apache.fory.graalvm.Main
+  run_graalvm_test org.apache.fory.graalvm.Main
 }
 
 graalvm_json_tests() {
-  run_graalvm_tests \
-    org.apache.fory.graalvm.ForyJsonExample \
-    org.apache.fory.graalvm.ForyJsonNoProviderExample
+  run_graalvm_test org.apache.fory.graalvm.ForyJsonExample
 }
 
 jdk25_access_options() {
