+++ integration_tests/graalvm_tests/README.md
@@ -1,9 +1,10 @@
 # GraalVM Native Image Tests
 
-Examples and tests for Fory serialization in GraalVM Native Image. The Fory JSON entry points are
-compiled with annotation processing disabled. They cover direct `JsonType` models, exact
+Examples and tests for Fory serialization in GraalVM Native Image. The Fory JSON entry point is
+compiled with annotation processing disabled. It covers direct `JsonType` models, exact
 `JsonMixin` target/source mappings, provider-selected hosted codec generation, configuration
-fallback to interpreted codecs, and a separate image with no reachable `ForyJsonProvider`.
+fallback to interpreted codecs, and hosted access metadata for unprovided configurations in one
+native image.
 
 ## Test
 
@@ -11,11 +12,6 @@ fallback to interpreted codecs, and a separate image with no reachable `ForyJson
 mvn -DmainClass=org.apache.fory.graalvm.ForyJsonExample clean -DskipTests=true -Dexec.skip=true -Pnative package
 ./target/main
 mvn -DmainClass=org.apache.fory.graalvm.ForyJsonExample clean -DskipTests=true -Pnative-module package
-./target/main-module
-
-mvn -DmainClass=org.apache.fory.graalvm.ForyJsonNoProviderExample clean -DskipTests=true -Dexec.skip=true -Pnative package
-./target/main
-mvn -DmainClass=org.apache.fory.graalvm.ForyJsonNoProviderExample clean -DskipTests=true -Pnative-module package
 ./target/main-module
 ```
